FasterEnum/requirements.txt
fpylll
flake8
begins
scipy
click
matplotlib
